Move `steps` under `common`
