+++ src/common/steps/Steps.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import withApplicationType from '../../common/withApplicationType/withApplicationType';
+import withApplicationType from '../withApplicationType/withApplicationType';
 
 import Step, { StepType } from './step/Step';
