+++ src/common/steps/step/step.scss
