+++ src/components/forms/sections/WinterOverviewInfo.tsx
@@ -14,7 +14,7 @@ import Person from '../fragments/overview/person/Person';
 import { ApplicationState } from '../../../redux/types';
 import { WinterFormValues } from '../../../types/winterStorage';
 import { WinterAreas } from '../../berths/types';
-import { StepType } from '../../steps/step/Step';
+import { StepType } from '../../../common/steps/step/Step';
 import { WithBoatType } from '../Selects';
 
 type Props = {
