+++ src/components/legends/berthLegend/BerthLegend.tsx
@@ -7,10 +7,10 @@ import Icon, { IconNames } from '../../../common/Icon';
 import AutoSave from '../../forms/AutoSave';
 import Form from '../../forms/Form';
 import ApplicationSelector from '../../forms/sections/applicationSelector/ApplicationSelector';
-import Steps from '../../steps/Steps';
+import Steps from '../../../common/steps/Steps';
 
 import { BerthsServices, SelectedServices } from '../../../types/services';
-import { StepType } from '../../steps/step/Step';
+import { StepType } from '../../../common/steps/step/Step';
 
 import './berthLegend.scss';
