+++ src/components/legends/winterStorageLegend/WinterStorageLegend.tsx
@@ -6,10 +6,10 @@ import { Col, Container, Row } from 'reactstrap';
 import Icon, { IconNames } from '../../../common/Icon';
 import AutoSave from '../../forms/AutoSave';
 import Form from '../../forms/Form';
-import Steps from '../../steps/Steps';
+import Steps from '../../../common/steps/Steps';
 
 import { SelectedWinterServices, WinterServices } from '../../../types/services';
-import { StepType } from '../../steps/step/Step';
+import { StepType } from '../../../common/steps/step/Step';
 
 import './winterStorageLegend.scss';
