+++ src/components/pages/selectedAreaPage/SelectedAreaPageContainer.tsx
@@ -23,7 +23,7 @@ import { Store } from '../../../redux/types';
 import { SelectedWinterServices } from '../../../types/services';
 import { WinterFormValues } from '../../../types/winterStorage';
 import { SelectedIds } from '../../berths/types';
-import { StepType } from '../../steps/step/Step';
+import { StepType } from '../../../common/steps/step/Step';
 import { Query } from 'react-apollo';
 
 interface Props {
